feat: Manage friend requests and notifications
This change introduces the ability to send, receive, and manage friend requests. It also includes UI updates to notify users of pending requests and a new `FriendRequest` type for data handling. The user data model has been updated to remove the `friends_uids` array, as this information is now managed through the friend request system.